FE: Delete `vue3-google-oauth2` package and use new GSI implementation
`vue3-google-oauth2` uses legacy method and Google doesn't allow that on newly created apps.

This commit makes `Sign in with Google` button on login page, but has no implementations for callback-ed credentials. To be implemented later
