+++ frontend/package.json
@@ -22,7 +22,6 @@
     "vue-cookies": "^1.8.1",
     "vue-property-decorator": "^9.1.2",
     "vue-router": "^4.1.3",
-    "vue3-google-oauth2": "^1.0.6",
     "vuex": "^4.0.2"
   },
   "devDependencies": {
