+++ frontend/src/pages/app/signin/LoginPage.vue
@@ -7,6 +7,7 @@
 <script lang="ts">
 import { Options, Vue } from "vue-class-component";
 import LoginWithGoogle from "@/components/signin/LoginWithGoogle.vue";
+import "@/plugins/signin/google/index";
 
 @Options({
   components: {
