--- frontend/src/plugins/signin/google-oauth2/config.ts
@@ -1,7 +1,0 @@
-import { GoogleAuthOptions } from "./interfaces";
-
-const options: GoogleAuthOptions = {
-  clientId: process.env.VUE_APP_SIGNIN_GOOGLE_CLIENT_ID,
-};
-
-export default options;
