--- frontend/src/plugins/signin/google-oauth2/index.d.ts
@@ -1,8 +1,0 @@
-declare module "vue3-google-oauth2" {
-  import { GoogleAuthOptions } from "./interfaces";
-
-  /**
-   * @default `options` = { scope: "profile email", prompt: "select_account" }
-   */
-  function install(app: any, options: GoogleAuthOptions);
-}
