+++ frontend/src/plugins/signin/google/index.ts
@@ -1,0 +1,6 @@
+const gsiScript = document.createElement("script");
+gsiScript.src = "https://accounts.google.com/gsi/client";
+gsiScript.async = true;
+gsiScript.defer = true;
+
+document.body.appendChild(gsiScript);
